+++ src/test/YamNet.Client.ConsoleTest/Tests/MessageTest.cs
@@ -31,15 +31,31 @@ namespace YamNet.Client.ConsoleTest
 
             Console.WriteLine("{0}. Get first five \"All\" messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetAll().Result.Take(5))
+            foreach (var message in client.Messages.GetAll().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;
             }
 
+            var references = client.Messages.GetAll().Result.References;
+            Console.WriteLine("\n{0}. Get a summary of the \"All\" messages references", j++);
+            i = 1;
+            Console.WriteLine("   1 - Messages: {0}", references.Count(q => q.Type == "message"));
+            Console.WriteLine("   2 - Tags    : {0}", references.Count(q => q.Type == "user"));
+            Console.WriteLine("   3 - Topics  : {0}", references.Count(q => q.Type == "topic"));
+            Console.WriteLine("   4 - Users   : {0}", references.Count(q => q.Type == "user"));
+
+            Console.WriteLine("\n{0}. Get first five of the \"All\" messages User references", j++);
+            i = 1;
+            foreach (UserReference user in references.Where(q => q.Type == "user").Take(5))
+            {
+                Console.WriteLine("   {0} - {1}", i, user.FullName);
+                i++;
+            }
+            
             Console.WriteLine("\n{0}. Get first five \"Feed\" messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetFeed().Result.Take(5))
+            foreach (var message in client.Messages.GetFeed().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;
@@ -47,7 +63,7 @@ namespace YamNet.Client.ConsoleTest
 
             Console.WriteLine("\n{0}. Get first five \"Top\" messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetTop().Result.Take(5))
+            foreach (var message in client.Messages.GetTop().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;
@@ -55,7 +71,7 @@ namespace YamNet.Client.ConsoleTest
 
             Console.WriteLine("\n{0}. Get first five \"Following\" messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetFollowing().Result.Take(5))
+            foreach (var message in client.Messages.GetFollowing().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;
@@ -63,7 +79,7 @@ namespace YamNet.Client.ConsoleTest
 
             Console.WriteLine("\n{0}. Get first five sent inbox messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetSent().Result.Take(5))
+            foreach (var message in client.Messages.GetSent().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;
@@ -71,7 +87,7 @@ namespace YamNet.Client.ConsoleTest
 
             Console.WriteLine("\n{0}. Get first five private inbox messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetPrivate().Result.Take(5))
+            foreach (var message in client.Messages.GetPrivate().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;
@@ -79,7 +95,7 @@ namespace YamNet.Client.ConsoleTest
 
             Console.WriteLine("\n{0}. Get first five inbox messages", j++);
             i = 1;
-            foreach (var message in client.Messages.GetReceived().Result.Take(5))
+            foreach (var message in client.Messages.GetReceived().Result.Messages.Take(5))
             {
                 Console.WriteLine("   {0} - {1}", i, message.MessageBody.Plain.ConsoleTrim(Chars));
                 i++;